REF: organize MultiIndex indexing tests (#31715)
